+++ indexer/secure_index_builder.go
@@ -1,4 +1,4 @@
-package secureIndexBuilder
+package indexer
 
 import (
 	"crypto/sha256"
